PrimaryPage to use BasePage
